requirements.txt
gitpython
gensim
scikit-learn
matplotlib
pygithub
tensorflow-gpu
beautifulsoup4
python-dotenv
pyyaml
pillow
pandas
numpy